backend/text_detect_ocr/requirements.txt
flask==2.3.3
opencv-python==4.8.1.78
numpy==1.24.3
werkzeug==2.3.7
omegaconf==2.3.0
onnxruntime==1.16.3
pyclipper==1.3.0
Pillow==10.0.1
torch==2.0.1
torchvision==0.15.2 